+++ app/Containers/Authentication/Configs/authentication-container.php
@@ -10,7 +10,7 @@ return [
     | When set to true, the user must confirm his email before being able to 
     | Login, after his registration.
     | 
-    */    
+    */
   
     'require_email_confirmation' => false,